+++ src/main/java/br/com/itau/desafio/acme/core/application/repository/InsuranceQuoteRepository.java
@@ -7,7 +7,7 @@ import java.util.UUID;
 
 public interface InsuranceQuoteRepository {
     void save(InsuranceQuote insuranceQuote);
-    Optional<InsuranceQuote> findById(Long id);
+    Optional<InsuranceQuote> findById(Integer id);
     Optional<InsuranceQuote> findByInsurancePolicyId(UUID insurancePolicyId);
     boolean existsByProductIdAndOfferIdAndCustomerDocument(UUID productId, UUID offerId, String documentNumber);
 }
